g3proxy: support ICAP reqmod for IMAP APPEND

+++ g3proxy/src/inspect/imap/authenticated.rs
@@ -33,7 +33,7 @@ enum ClientAction {
     Loop,
     Logout,
     Idle,
-    SendLiteral(usize),
+    SendLiteral(u64),
 }
 
 pub(super) enum CloseReason {
