+++ lib/g3-imap-proto/src/response/mod.rs
@@ -96,7 +96,7 @@ pub enum CommandData {
 
 pub struct UntaggedResponse {
     pub command_data: CommandData,
-    pub literal_data: Option<usize>,
+    pub literal_data: Option<u64>,
 }
 
 impl UntaggedResponse {
@@ -303,11 +303,11 @@ impl Response {
     }
 }
 
-fn check_literal_size(left: &[u8]) -> Result<Option<usize>, ResponseLineError> {
+fn check_literal_size(left: &[u8]) -> Result<Option<u64>, ResponseLineError> {
     if left.ends_with(b"}") {
         if let Some(p) = memchr::memrchr(b'{', left) {
             let size_s = &left[p + 1..left.len() - 1];
-            let (size, offset) = usize::from_radix_10_checked(size_s);
+            let (size, offset) = u64::from_radix_10_checked(size_s);
             if offset != size_s.len() {
                 return Err(ResponseLineError::InvalidLiteralSize);
             }
